Remove deprecated UI bundle workflow
Deleted the `build-bundle.yml` workflow, which was responsible for building and publishing the UI bundle to releases. This workflow is now obsolete due to a shift in the build and deployment strategy. It is no longer needed to automate UI bundle releases in this manner.

Release-Note: remove the deprecated UI bundle GitHub Actions workflow

Signed-off-by: Ian Skelskey <ianskelskey@gmail.com>